Implement Gf128 struct with basic operations and tests; update bits and portable modules

+++ src/bits.rs
@@ -23,7 +23,7 @@ pub fn toggle_bit(x: u64, i: usize) -> u64 {
 pub fn to_bits(x: u64) -> [bool; 64] {
     let mut v = [false; 64];
     for i in 0..64 {
-       v[i as usize] = get_bit(x, i);        
+       v[i] = get_bit(x, i);        
     }
     v
 }
@@ -31,7 +31,7 @@ pub fn to_bits(x: u64) -> [bool; 64] {
 pub fn from_bits(bits: [bool; 64]) -> u64 {
     let mut x: u64 = 0;
     for i in 0..64 {
-    x = set_bit(x,i,bits[i as usize]);
+    x = set_bit(x,i,bits[i]);
     }
     x
 }
